Update ui_vscript.gnut
I'm not sure what this file is for.
I had added that line of code at the beginning because `MenuMapSelect_Init()` was there, and I was using this menu as a reference to create the map select menu for the server browser.
However, after testing, I found that the menu works perfectly fine without that line, so I decided to remove it.