Change function to recieve index of Titan
`GetPersistentSpawnLoadoutIndex` directly in the function is returning wrong Titan index sometimes, and will reset an unwanted Titan instead. Changing the function to recieve the Titan idex from `uiGlobal.titanSpawnLoadoutIndex` seems to work better, this method also allows calling the command from console to specify a Titan to reset that is not the current selected one.