Fix players dying when mantling on props or embarking Titans in `mp_complex3` (#709)
There's a whole `trigger_hurt` right below the ground in the inner large hallway which whenever a player tries to mantle on any of the props in there either intentionally or by accident they will instantly die. This also applies to when you embark your titan in there. This does not happen on vanilla servers.

This commit removes the trigger that kills the player.